Circle fixes: extract ApolloWrapper component
